backend/pyproject.toml

[tool.poetry]
name = "readpilot-backend"
version = "0.1.0"
description = "AI-powered reading companion backend"
authors = ["SuperYing <13975145507@163.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.0"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
aiosqlite = "^0.21.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"
python-multipart = "^0.0.20"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.28.0"
celery = "^5.4.0"
redis = "^5.2.0"
pymupdf = "^1.25.3"
ebooklib = "^0.18.0"
python-docx = "^1.1.2"
markdown = "^3.7.0"
langchain-text-splitters = "^0.3.2"
openai = "^1.59.0"
anthropic = "^0.42.0"
chromadb = "^0.5.0"
tenacity = "^9.0.0"
pillow = "^11.1.0"
beautifulsoup4 = "^4.12.0"
email-validator = "^2.3.0"
greenlet = "^3.2.4"
bcrypt = ">=5.0.0"
structlog = "^25.4.0"
dashscope = "^1.24.7"
aiofiles = "^25.1.0"
boto3 = "^1.40.57"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4"
pytest-asyncio = "^1.0"
ruff = "^0.8"
mypy = "^1.13"

[tool.poetry.scripts]
dev = "scripts:dev"
prod = "scripts:prod"
test = "scripts:test"
lint = "scripts:lint"
format = "scripts:format"
db-init = "scripts:db_init"
db-migrate = "scripts:db_migrate"
db-upgrade = "scripts:db_upgrade"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["app/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = ["-v", "--strict-markers", "--tb=short"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
